+++ component/frontend/Model/Subscribe.php
@@ -7,7 +7,6 @@
 
 namespace Akeeba\Subscriptions\Site\Model;
 
-use Akeeba\Subscriptions\Admin\Helper\Forex;
 use Akeeba\Subscriptions\Admin\PluginAbstracts\AkpaymentBase;
 use Akeeba\Subscriptions\Site\Model\Subscribe\StateData;
 use Akeeba\Subscriptions\Site\Model\Subscribe\Validation;
@@ -913,16 +912,6 @@ class Subscribe extends Model
 			// Flags
 			'_dontCheckPaymentID'        => true,
 		);
-
-		// If I have an alternate currenct, let's convert some values and store them
-		$currency_alt  = $this->container->params->get('invoice_altcurrency', '');
-
-		if($currency_alt)
-		{
-			$currency = $this->container->params->get('currency', 'EUR');
-
-			Forex::updateRates(false, $this->container);
-		}
 
 		/** @var Subscriptions $subscription */
 		$subscription = $this->container->factory->model('Subscriptions')->tmpInstance();