+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Removed the Invoice Templates feature (gh-345)
 * Removed the Credit Notes Templates feature (gh-347)
 * Removed all tax-related features (gh-334)
+* Removed all currency exchange rate features (gh-356)
 
 # 6.2.3